Move tests into separate directories
