+++ Tests/SQLiteTests/Core/ConnectionTests.swift
@@ -111,11 +111,6 @@ class ConnectionTests: SQLiteTestCase {
         XCTAssertEqual(2, db.totalChanges)
     }
 
-    func test_userVersion() {
-        db.userVersion = 2
-        XCTAssertEqual(2, db.userVersion!)
-    }
-
     func test_prepare_preparesAndReturnsStatements() throws {
         _ = try db.prepare("SELECT * FROM users WHERE admin = 0")
         _ = try db.prepare("SELECT * FROM users WHERE admin = ?", 0)
@@ -444,52 +439,5 @@ class ConnectionTests: SQLiteTestCase {
             }
         }
         semaphores.forEach { $0.wait() }
-    }
-
-    func test_attach_detach_memory_database() throws {
-        let schemaName = "test"
-
-        try db.attach(.inMemory, as: schemaName)
-
-        let table = Table("attached_users", database: schemaName)
-        let name = Expression<String>("string")
-
-        // create a table, insert some data
-        try db.run(table.create { builder in
-            builder.column(name)
-        })
-        _ = try db.run(table.insert(name <- "test"))
-
-        // query data
-        let rows = try db.prepare(table.select(name)).map { $0[name] }
-        XCTAssertEqual(["test"], rows)
-
-        try db.detach(schemaName)
-    }
-
-    func test_attach_detach_file_database() throws {
-        let schemaName = "test"
-        let testDb = fixture("test", withExtension: "sqlite")
-
-        try db.attach(.uri(testDb, parameters: [.mode(.readOnly)]), as: schemaName)
-
-        let table = Table("tests", database: schemaName)
-        let email = Expression<String>("email")
-
-        let rows = try db.prepare(table.select(email)).map { $0[email] }
-        XCTAssertEqual(["foo@bar.com"], rows)
-
-        try db.detach(schemaName)
-    }
-
-    func test_detach_invalid_schema_name_errors_with_no_such_database() throws {
-        XCTAssertThrowsError(try db.detach("no-exist")) { error in
-            if case let Result.error(message, code, _) = error {
-                XCTAssertEqual(code, SQLITE_ERROR)
-                XCTAssertEqual("no such database: no-exist", message)
-            } else {
-                XCTFail("unexpected error: \(error)")
-            }
-        }
     }
 }
