+++ Tests/SQLiteTests/Core/CoreFunctionsTests.swift
