+++ Tests/SQLiteTests/Core/ResultTests.swift
