+++ Tests/SQLiteTests/Typed/AggregateFunctionsTests.swift
