+++ Tests/SQLiteTests/Typed/CustomAggregationTests.swift
