+++ Tests/SQLiteTests/Typed/CustomFunctionsTests.swift
