+++ Tests/SQLiteTests/Typed/OperatorsTests.swift
